support wheel scroll Wrap the component in a div with an onWheel handler.
