+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SingleImageViewer/SingleImageViewerContainer.spec.js
@@ -232,64 +232,4 @@ describe('Component > SingleImageViewerContainer', function () {
       })
     })
   })
-
-  describe('with pan and zoom', function () {
-    const onError = sinon.stub()
-    const onReady = sinon.stub()
-    let onPan
-    let onZoom
-
-    beforeEach(function () {
-      const subject = {
-        id: 'test',
-        locations: [
-          { 'image/jpeg': 'https://some.domain/image.jpg' }
-        ]
-      }
-      wrapper = shallow(
-        <SingleImageViewerContainer
-          ImageObject={ValidImage}
-          subject={subject}
-          onError={onError}
-          onReady={onReady}
-          setOnPan={callback => { onPan = callback }}
-          setOnZoom={callback => { onZoom = callback }}
-        />
-      )
-      wrapper.instance().imageViewer = {
-        current: {
-          clientHeight: 50,
-          clientWidth: 100,
-          addEventListener: sinon.stub(),
-          getBoundingClientRect: sinon.stub().callsFake(() => ({ width: 100, height: 50 })),
-          removeEventListener: sinon.stub()
-        }
-      }
-    })
-
-    it('should should zoom out on wheel scroll up', function (done) {
-      onReady.callsFake(function () {
-        const { onWheel } = wrapper.instance()
-        onWheel({ deltaY: 10, preventDefault: sinon.stub() })
-        let viewBox = wrapper.find(SingleImageViewer).prop('viewBox')
-        expect(viewBox).to.equal('18.5 9.5 363 181')
-        onWheel({ deltaY: -10, preventDefault: sinon.stub() })
-        viewBox = wrapper.find(SingleImageViewer).prop('viewBox')
-        expect(viewBox).to.equal('0 0 400 200')
-        done()
-      })
-    })
-
-    it('should should zoom in on wheel scroll down', function (done) {
-      onReady.callsFake(function () {
-        const { onWheel } = wrapper.instance()
-        let viewBox = wrapper.find(SingleImageViewer).prop('viewBox')
-        expect(viewBox).to.equal('0 0 400 200')
-        onWheel({ deltaY: 10, preventDefault: sinon.stub() })
-        viewBox = wrapper.find(SingleImageViewer).prop('viewBox')
-        expect(viewBox).to.equal('18.5 9.5 363 181')
-        done()
-      })
-    })
-  })
 })
